codicons: Fix empty derived list
