+++ src/vs/base/common/codiconsLibrary.ts
@@ -8,7 +8,7 @@ import { register } from 'vs/base/common/codiconsUtil';
 // This list is automatically generated by the vscode-codicons repo.
 // Please don't edit it, as your changes will be overwritten.
 // If you want to create a mapping, add it to codiconsExtras.ts
-export const codiconsCore = {
+export const codiconsLibrary = {
 	add: register('add', 0xea60),
 	plus: register('plus', 0xea60),
 	gistNew: register('gist-new', 0xea60),
